% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BuchDarrah19}
\alias{BuchDarrah19}
\title{Example data for gretel}
\format{a numeric matrix with 5 rows and 5 columns}
\source{
<DOI:10.1016/j.socnet.2010.03.006>
}
\usage{
BuchDarrah19
}
\description{
A sociomatrix encoding tie strengths among five nodes
}
\keyword{datasets}
