% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dijkstra_inf}
\alias{dijkstra_inf}
\title{Find the shortest L-Inf norm paths to other vertices}
\usage{
dijkstra_inf(dist, src)
}
\arguments{
\item{dist}{A matrix of distances between nodes}

\item{src}{An integer vertex ID}
}
\value{
A numeric vector, entry \emph{i} of which is the vertex immediately preceeding 
  vertex \emph{i} in the shortest path leading to \emph{i}. Full paths must be constructed
  recursively.
}
\description{
Find the shortest L-Inf norm paths to other vertices
}
