% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dijkstra_nodes}
\alias{dijkstra_nodes}
\title{Find the shortest paths to other vertices}
\usage{
dijkstra_nodes(dist, src, node_costs)
}
\arguments{
\item{dist}{A matrix of distances between nodes}

\item{src}{An integer vertex ID}

\item{node_costs}{a list of costs, in order, of all nodes represented in the 
sociomatrix, all are assumed 0 if unspecified}
}
\value{
A numeric vector, entry \emph{i} of which is the vertex immediately preceeding 
    vertex \emph{i} in the shortest path leading to \emph{i}. Full paths must be constructed
    recursively.
}
\description{
Find the shortest paths to other vertices
}
