% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OpsahlEtAl10}
\alias{OpsahlEtAl10}
\title{Example data from Opsahl, Agneessens, Skvoretz (2010)}
\format{a numeric matrix with 5 rows and 5 columns}
\source{
<DOI:10.1016/j.socnet.2010.03.006>
}
\usage{
OpsahlEtAl10
}
\description{
A sociomatrix encoding tie strengths among five nodes, used for examples in
Opsahl, Agneessens, Skvoretz (2010) Social Networks 32(2010):245-251
}
\keyword{datasets}
