% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{unpack}
\alias{unpack}
\title{Unpacks a Path from a Dijkstra-Format Spanning Tree}
\usage{
unpack(tree, source, target)
}
\arguments{
\item{tree}{a Dijkstra-format tree returned by \code{all_opt_gpv} or \code{all_opt_ppv}}

\item{source}{an integer index corresponding to a node in \code{sociomatrix}}

\item{target}{an integer index corresponding to a node in \code{sociomatrix}}
}
\description{
Used with \code{all_opt_gpv} and \code{all_opt_ppv} to 
unpack individual paths from the Dijkstra-format trees that 
those functions return.
}
\details{
Returns \code{NA} if a path does not exist
}
