% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{YangKnoke01}
\alias{YangKnoke01}
\title{Example data from Yang, Knoke (2001)}
\format{a numeric matrix with 5 rows and 5 columns}
\source{
<DOI: 10.1016/S0378-8733(01)00043-0>
}
\usage{
YangKnoke01
}
\description{
A sociomatrix encoding tie strengths among five nodes, used 
for examples in Yang, S., Knoke, D. (2001) Social Networks 23(4):285-295
}
\keyword{datasets}
